+++ compiler/src/codegen/compile_time.rs
@@ -581,7 +581,7 @@ fn render_sort(sort: &Sort) -> String {
         Sort::Effect => "effect".to_owned(),
         Sort::Effects => "effects".to_owned(),
         Sort::Parameters => "parameters".to_owned(),
-        Sort::Fragment(crate::ast::StaticFragmentKind::Constraint) => "constraint".to_owned(),
+        Sort::Fragment(kind) => kind.as_str().to_owned(),
         Sort::ParameterPack => "...parameters".to_owned(),
         Sort::ParameterModifier => "(P: parameters): parameters".to_owned(),
         Sort::TypeConstructor { parameter_groups } => {